Make ListenerAdapter serialize JSON directly
Change websocket notification serialization to serialize
to json directly instead of first serializing to xml and
subsequently converting to json.
While we are here fixup the websocket notification format
so it actually conforms to the sal-remote.yang model and
encode namespaces according to the json-ietf standard
instead of bundling an xmlns node in the json ouptut.

JIRA: NETCONF-471
Change-Id: I1b90c99980c0692b217523640bd6f67050f21e14
Signed-off-by: Tomas Cere <tomas.cere@pantheon.tech>