Update patch set 19
Patch Set 19: Code-Review-1

(6 comments)

Couple of minor issues, but otherwise seems ok to me.
Also in the future when doing something of this scope, I would split this into two patches, one would be the new api and implementation on the netconf side of the world and the other would be the restconf changes that would pick it up and use it. It makes review easier ;) but its fine this time no need to split it up now.

Patch-set: 19
Reviewer: Gerrit User 3127 <3127@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1