Add import for netconf-node-optional
This was missed in the initial work to add the
optional schema first added here:
  https://git.opendaylight.org/gerrit/c/netconf/+/82521

Because of that, adding any of the optional configs
(e.g., datastore-lock) to the body of the create device
RPC was failing. importing it in netconf-node-topology
fixes that.

JIRA: NETCONF-670
Change-Id: I5fc72d9714e0fb4f1dfd26f161d2eb3dcfe8862a
Signed-off-by: Nikhil Soni <nsoni@luminanetworks.com>
(cherry picked from commit 9ed4befdde4f9c88705ddd1ec70c5b5b5ec19a25)