Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > This needs an explanation and probably a Jira issue
> 
> This is literally only about making a header value more readable. turning something like
> this:
> 'ETag': '"2017-01-26streams"
> 
> in to this:
> 
> 'ETag': '"2017-01-26-streams"
> 
> Kailash, maybe make a quick netconf jira about making ETag header more readable? then update the
> commit message to say what I just wrote above and link the jira to it.

Ah, now I understand. No need for a Jira, but the commit message needs to be rephrased -- as it currently stands it makes it sounds as though we are violating some spec or something (which lead me on a wild goose chase).

Patch-set: 3
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>