Do not fail on warnings for docs-linkcheck
This is a workaround due to sphinx-tabs being incompatible with the
Sphinx linkchecker which throws a warning. We mainly use linkchecker
to catch broken links so we don't really need to fail on warnings as
the main "docs" builder will still fail for warning level issues so
this is a reasonable workaround.

This patch also fixes some current broken links in the docs, mainly
wiki.odl to wiki-archive.odl as these docs have not been migrated
to the new wiki as of yet.

Signed-off-by: Thanh Ha <zxiiro@gmail.com>
Change-Id: I94d09ad6740efde204b4bb649623eedf346c3208