Filter structure for XPath in Netconf message transformer
Creating implementation for XPath type filter element in XML body
of request which allows to send get and get-config operations to
read data from Netconf device with use of the XPath. If the device
doesn't support the XPath, the XPath will be transformed and wrote
into the filter element as the subtree.

Change-Id: I845e06831be4639e74aee31a1e51b145f697a1a8
Signed-off-by: Jakub Toth <jakubt4@gmail.com>