Modularize the NETCONF user guide
These files exist in the current NETCONF user guide, except
they are all in one file (user-guide.rst). I took each level 1
head and made it a separate .rst file
In addition, I rewrote the documents for clarity and to correct
some errors. Also, I updated some commands so that they follow
the same format for consistancy. For example, I changed the
GET http://localhost:8181/restconf/operational/network-topology:network-topology/
command so that it matched the command just above this command.

Still, if you start with the index, the flow of the document
is the same as the old document.

Signed-off-by: David Baugher <dbaugher@luminanetworks.com>
Change-Id: I62227ba194cf5dbfb586737707070b064097a411