Eliminate unnecessary blocking checks
According to RFC6241:
1) "delete" operation must first check if the configuration data
   currently exists in the configuration datastore.
2) "create" operation must first check if the configuration data does
   not already exist in the configuration datastore.
3) "replace" operation replaces any related configuration in the
   configuration datastore. If no such configuration data exists in
   the configuration datastore, it is created.

If request goes via NetconfRestconfStrategy, we do not need to check
existence on the controller side, as the implied checks are part of
the underlyig behavior.

If request goes via MdsalRestconfStrategy, first we need to do
"if exists" check. Read requests in this case are collected in a batch
and then collecting the results.

JIRA: NETCONF-403
Change-Id: I6354c345d099017c1e82878af36ac37bbf5662c9
Signed-off-by: Vladyslav Marchenko <vladyslav.marchenko@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>