Eliminate unnecessary blocking checks
According to RFC6241:
1) "delete" operation must first check if the configuration data currently exists in the configuration datastore.
2) "create" operation must first check if the configuration data does not already exist in the configuration datastore.
3) "replace" operation replaces any related configuration in the configuration datastore. If no such configuration data exists in the configuration datastore, it is created.

All these checks are done for NetconfRestconfStrategy flow (in org.opendaylight.netconf.mdsal.connector.ops.EditConfig.java). But not for MdsalRestconfStrategy. So for NetconfRestconfStrategy we did this check twice. Also changed the logic to first check for presence and then perform modifications. And also allows batching of modification requests.

JIRA: NETCONF-403
Change-Id: I6354c345d099017c1e82878af36ac37bbf5662c9
Signed-off-by: Vladyslav Marchenko <vladyslav.marchenko@pantheon.tech>