Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> So we have a direct netconf/restconf API in NetconfDataTreeService, why can't we just use that instead of bothering with DOMDataBroker?

We have applications built on top of NETCONF that are using just transaction-based API (simplified usability is a main point since we would have to use combination of two services). From the view of RESTCONF it is, of course, useless - so I can remove it in this context.

Patch-set: 7