WIP: Fix shaded sshd
sshd-2.5.0 introduced a regression in that it is not closing
sessions correctly. This was reported upstream in
https://issues.apache.org/jira/browse/SSHD-1028, but since there
is no release yet, we need to apply a workaround.

This patch does so by introducing a bit of sneakiness: it uses
a pre-compiled version of Nio2Session instead of upstream.

// FIXME: add an explicit test

JIRA: NETCONF-736
Change-Id: I23c52616f74033df99f271a6a940271fa9f96b36
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>