Do not use org.opendaylight.controller.sal.common.util.NoopAutoCloseable
NoopAutoCloseable hards back to Java 7 days and is an anachronism
with advent of lambdas. Replace its use with no-op lambdas.

As this is the last class used from sal-common-util in both components,
remove that dependency as well.

A quick follow-up audit shows we can remove the dependency from two
other artifacts, leaving rest-nb-bierman02 the only artifact still
needing it for AbstractMXBean.

Change-Id: Id97d3cdf110047778248e21021db39a6da2c3a5b
Signed-off-by: Vladyslav Marchenko <vladyslav.marchenko@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>