Update patch set 3
Patch Set 3:

(4 comments)

> Patch Set 2: Code-Review-1
> 
> Okay, so this does a metric ton of operations. Why exactly do we need this? We are very much in compliance with RFC AFAICT without doing all this.

"so this does a metric ton of operations" - There is still a single GET RPC. From the view of CPU processing, there is only single addition that compares key definitions against actual list keys (there aren't usually many keys). This modification doesn't affect previous implementation at all, only fields-aware filtering.

"Why exactly do we need this?" - we send, for example, this RPC with following filter (it is perfectly valid case):

        <filter xmlns:ns0="urn:ietf:params:xml:ns:netconf:base:1.0" ns0:type="subtree">
            <root xmlns="urn:ietf:params:xml:ns:yang:test-model">
                <list-root>
                    <top-list>                      
                        <next-data>
                            <switch-2/>
                        </next-data>
                    </top-list>
                </list-root>
            </root>
        </filter>

"top-list" is described by list schema node. And device responds with such data:

<top-list>
  <next-data>
    <switch-2>true</switch-2>
  </next-data>
</top-list>
<top-list>
  <next-data>
    <switch-2>false</switch-2>
  </next-data>
</top-list>
...

The point is that NETCONF server doesn't have to send us keys - and this behaviour is also valid according to RFC (some versions of ConfD behaves in this way). And without keys, NETCONF client could not parse the response.

Possible solutions:
a. we always ask for key-leaves - NETCONF server is responsible for parsing of response, so why it should not ask for these keys, if it needs them?
b. creation of custom XML parser that will not need list keys
c. user will be responsible for asking for list keys - it will decrease usability, in the real world of customers it is not acceptable

I chose the approach (a)

Patch-set: 3