Netconf lock/unlock issue when multiple request arrives for the same device
When device supports both the capability :candidate and :writable-running.
First we execute lock and all edit-config operation. If there are no errors, only then we do commit operation.
If there are some errors in any of the previous operations - discarding all the changes.
Before we have the situation that we do commit even if we have errors in previous operations.

JIRA: NETCONF-705
Change-Id: I37d046217581d531cc8d31ed4542486b36d7798b
Signed-off-by: Vladyslav Marchenko <vladyslav.marchenko@pantheon.tech>