Netconf lock/unlock issue when multiple request arrives for the same device
First we execute lock operation. If we fail to lock - non of the further operations will be executed.
If lock is succeeded - all edit-config operations will be executed. If there are no errors, only then we do commit and unlock operations.
If there are some errors in any of the previous operations before commit - discarding all the changes.
Before we had the situation that we do commit even if we have errors in previous operations.

JIRA: NETCONF-705
Change-Id: I37d046217581d531cc8d31ed4542486b36d7798b
Signed-off-by: Vladyslav Marchenko <vladyslav.marchenko@pantheon.tech>