YANG patch request on NETCONF mountpoint is not atomic
First we execute lock operation. If we fail to lock - non of the further operations will be executed.
If lock is succeeded - all edit-config operations will be executed. If there are no errors, only then we do commit and unlock operations.
If there are some errors in any of the previous operations before commit - discarding all the changes.
Running datastore is not supported discard operation. In this case snapshot of database is created. If we need to rollback all "edit-config" changes, this snapshot is used in "copy-config" operation. In this way we will rollback the database to the state before the lock operation.
This issue is depends on NETCONF-705

JIRA: NETCONF-746
Change-Id: I4d091b566ceba5ce75aa61da253e7bbb21728732
Signed-off-by: Vladyslav Marchenko <vladyslav.marchenko@pantheon.tech>