Deprecate JSONRestconfService
This service is type-unsafe and does not have users in OpenDaylight
proper. The contract is actively hurtful as it seeks to steer users
towards opaque blobs, not expressing their service and data
dependencies. Furthermore the implementation structure forces an ugly
interface on the actual RESTCONF services, preventing reasonable
refactors.

Deprecate the service for removal, so that we can get rid of this burden
in the next major release.

Change-Id: I96d3fd86ba773fa8986454d1ac4825723b15152d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>