Update patch set 28
Patch Set 28:

I have checked it from functional perspective and did not find anything that needs to be fixed, except only for case when --generate-configs-batch-size is larger than --device-count, it raise an Exception, but works, creates and connects --generate-configs-batch-size number of devices. So it is not necessary to fix it as it works even in such case.

Patch-set: 28
Reviewer: Gerrit User 8907 <8907@7fc14799-209e-464c-9743-7a06c2c21a81>