Eliminate use of AsyncSshHandlerReader in call home
Using AsyncSshHandlerReader did not require ChannelHandlerContext for reading.
After eliminating AsyncSshHandlerReader, we need to create a ChannelSubsystem
with the ChannelHandlerContext, which we can only obtain after
creating and calling the open listener on the ChannelSubsystem.

Current solution is to utilize pipelines from MinaSshNettyChannel.
The logic from MinaSshNettyChannel has been moved to CallHomeSessionContext,
where a ChannelOutboundHandlerAdapter is created and added as the first
in the pipeline. This call will create the ChannelHandlerContext
in the pipeline, and it can be used later inside the ChannelSubsystem.

JIRA: NETCONF-699
Change-Id: I58c702949bc31061b89b12d166f0fa6b8bc383c7
Signed-off-by: Kostiantyn Nosach <kostiantyn.nosach@pantheon.tech>
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>