Remove RestconfError.ErrorTag
The concept of an error-tag is shared between NETCONF and RESTCONF, or
anything iteracting with YANG. We have a unified definition in
yang.common, which NETCONF is already using. Eliminate
RestconfError.ErrorTag in favor of the yang.common.

Implementation updates are rather straighforward, except the fact we end
up having custom well-known error-tags. Deal with this through
relocation towards RESTCONF-specific ErrorTags, which concentrate all
related tasks.

JIRA: NETCONF-793
Change-Id: I21a79625a2509adabf50f706739c88f9a3359777
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>