Do not fail on actions without output
During the investigation of NETCONF-794
we have found that Restconf throws exception
when there is no data returned by action invocation.

This is not correct because actions without output
are legal.

We have fixed this according to RFC 8040 section 6.3
so in case action returns no result data we return
empty body with status code 204.

Change-Id: Ie8f3d098fc65828998b088bfed5ca89e2f438828
Signed-off-by: PeterSuna <Peter.Suna@pantheon.tech>
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>