Fix incorrect handling of list path in the RESTCONF fields
emitChildTreeNode for AbstractMapMixin fails because it's called for
the children of subtree instead of the subtree.
Also, there is a mismatch when handling a list (AbstractMapMixin)
because the getchildren method returns the inner node (ListEntry)
while the subtree has the data nodes inside the list.

Create streamToWriter for AbstractMapMixin, and other classes with
inner node, and properly handle key leaves.

JIRA: NETCONF-820
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>
Change-Id: I23ff3a8aadfa19f577bcb7e665f3b029fd6c6d96