Add rfc8040.ApiPath
Our YangInstanceIdentifierSerializer is a rather complex piece of
machinery, which is a maintenance burden and cannot be modified easily.

In order to deal with it, we will split parsing into two steps:

1) structural tokenization, which splits the input string into a
   representation of 'api-path' ABNF producation.

2) semantic binding to an EffectiveModelContext, which reconciles the
   structure with the context, producing YangInstanceIdentifier in the
   process.

This patch deals with item 1) by introducing an ApiPath class, which
can be constructed from a string.

// FIXME: deal with percent-encoded /,

JIRA: NETCONF-631
Change-Id: I44c8ad1308b27ee459f95f8617f54c8537ea9c65
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>