WIP: Add ApiSegment
// FIXME: this needs a bit more thought, as a direct conversion to
//        YangInstanceIdentifier + SchemaInferenceStack is a tad
//        problematic

The crux of the issue do not know how the path is being interpreted --
i.e. we need to drive it from each access method. The reason for that is
that sometimes the entry refers to a pure data resource, while at other
times it can refer to an operation resource.

We can detect this quite easily through SchemaInferenceStack, but there
is a a slight inefficiency in that an attempt to enter data tree for an
action (or RPC) would entail throwing an exception -- which is slow.

// FIXME: finish implementation
// FIXME: plug into parsing machinery

JIRA: NETCONF-718
Change-Id: I81add7b615e90da984a89e5d6faee81b44043e8c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>