Fix RestconfOperationsService.getOperations(UriInfo)
ietf-restconf.yang has a rather ugly wart in its definition of the
operations container, which we are using an ugly workaround for.

This workaround is rendered inoperable due to strict binding to
effective model context, which does not find the dynamic leaves required
to encode the content.

Rework the implementation to side-step NormalizedNode documents and
provide ready JSON/XML strings.

JIRA: NETCONF-822
Change-Id: I57760d06240e09940026fee9b195e207853c01b2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>