Improve YangInstanceIdentifierDeserializer.checkValid()
The current implementation is slow due to always playing with arguments,
even when we do not use them.

As it turns out, a little bit of specialization goes a long way, as we
have only two types of callers: 0 and 1 argument. Take that into
account, further cleaning things up by inlining the format string
fragmet.

Change-Id: Ic5de0d473fe203703470252b32e54b2e3c325f5e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>