Replace entity-owners API call
An entity ownership test for OpenFlow Plugin makes a direct REST
API call to get the entity ownership information from operational
data store, which is only valid up to Silicon.

Update the 'get_entities' function in the ClusterEntities library
to fall back on the entity-owners API call for Silicon, and replace
the API call to the operational data currently used in the test case
with the function/keyword so that Silicon/Phosphorus/Sulfur are
all supported.

Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>
Change-Id: Idf7d41d1abc509f5f344e7d6d40d93a4dee8f960