Fix incorrect level assignment during fields parsing
Before, everytime a sub-selector was encountered, new level would be created.
This is unwanted behaviour and caused missing data in RESTCONF and for NETCONF
mountpoints it would fail with NPE because LinkedPathElement was expected on previous
level but could not be found.
Add proper level tracking and test-cases.

JIRA: NETCONF-660
Signed-off-by: Peter Puškár <ppuskar@frinx.io>
Change-Id: Icaf7c181b23ce9f7110cd6f63f17b55b4c1bc322