Refactor RestDocgenUtil
To compare name-space and revision we work with QNames -
we have adapted method resolvePathArgumentsName(..) parameters
to require QNames.

We can go further and mark all parameters as NonNull because
the use-case when parent is null has no sense - we are at direct
child of module - so full path should be used.

Align parameter usage across the methods -
first is node's QName the second is node's parent's QName.

Mark not-used resolveNodesName(SchemaNode, Module) method
as deprecated for removal.

JIRA: NETCONF-819
Change-Id: I64c5317af45ce4d0d182123922e71ab5c9d1c0c0
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>