WIP: Fix notification serialization when using odl-leaf-nodes-only query parameter
Fix SchemaPath path for node serialization, augmentation nodes do not have QName, so we need to filter them out as well.
Add tests to cover serialization of augmentation and choice nodes.
This patch needs to wait for fix in YANGTOOLS because currently we cannot initialize Nested writers (both JSON and XML)
on choice nodes.
After this is fixed, enable choice nodes in tests.

JIRA: NETCONF-848
Change-Id: I9f6dffd9532bd3d52021bbac58473c3cf3b58937
Signed-off-by: Peter Puškár <ppuskar@frinx.io>