Fix periodic NETCONF Call Home connection dropping
Callhome devices make reconnection every hour due to key re-exchange, which
is part of SSHD implementation.

When a specific session is authenticated and activated, in case of key
re-exchange, authentication (invoking of doAuth() method) does not need to
be made again.

While we are at it, also make sure CallHomeSessionContext's constructor
does not have side-effects -- while this cannot quite happen since we do
not reuse ClientSessions, it is a needless leak.

JIRA: NETCONF-681
Change-Id: I824c92d230c7570570d5eed21d489c435bdc8b22
Signed-off-by: Ivan Martiniak <ivan.martiniak@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>