Customize NetconfEMOAggregator decode method
Delimiter search in DelimiterBasedFrameDecoder becomes the bottleneck
if decode() is repeatedly called for small increments of a large frame.

Create NetconfEOMAggregator directly from ByteToMessageDecoder, instead of
extending DelimiterBasedFrameDecoder, and implement a custom decode() method
customized for NETCONF end-of-message framing.
The implemented decoding method keeps track of the progress in the input buffer to avoid repeated scanning of delimiter from readerIndex.

JIRA: NETCONF-889
Change-Id: I40a9ccd0f7a68356098aa95cb2e2b4b2b95a2b53
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>