Make max chunk size property configurable
Netconf client with default max chunk size value(16MB) is unable to
parse the yang files from some of the devices.

By increasing the max chunk size value to 32MB,
the netconf client is able to parse those yang files.

In the existing codebase, the default max chunk size value is
hard-coded. And the max chunk size value needs to be changed
based on the devices, it will be good if we keep this property
configurable.

JIRA: NETCONF-888
Signed-off-by: DanielCV <daniel.viswanathan@verizon.com>
Change-Id: Iec041a4ba9c8886cceb44fa86d07320bb5ae942b