Add {keystore,truststore}-api and none implementations
The configuration bits for NETCONF/RESTCONF clients and server are part
of a large scale effort in NETCONF WG. In particular the concept of
crypto types and a keystore/truststore are shared between SSH and TLS
transports.

This patch adds the YANG models and an a 'none' implementation, which
the central keystore/truststore.

JIRA: NETCONF-590
Change-Id: I0827756df83aa9cfc83e98ee50b1e44f6310e82e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>