Create RemoteDeviceServices.{Actions,Rpcs}
Provide proper encapsulation of action and RPC invocation strategies. We
do not define schema-less actions, as those cannot be supported, but
keep our options open for future.

This flushes out the need to have SchemalessRpcService to cover generic
schema-less use case, but also NetconfRpcService, which works on base
NETCONF rpcs -- and that is used by most of sal-netconf-connector.

Also update tests to use Futures instead of FluentFutures and simplify
any(Class) to any(), as we have accurate types already.

JIRA: NETCONF-930
Change-Id: I75f1cb67f50a295bbce6638009d663444a48ca9c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>