Do not touch config datastore during initiliazation
We should not be introducing configuration items in the datastore, as
that places onus on operator and/or the datastore to deal with this
magic occuring. This also removes a cross-datastore transtion -- which
is no longer supported in Argon.

This means the configuration datastore is empty on startup, hence tests
need to initialize it. This also provides a stepping stone towards
allowing multiple netconf-capable topologies to exist.

JIRA: NETCONF-919
Change-Id: I0218b19f37dbd9ccd3196040193f305d5d147a89
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>