Simplify NetconfDeviceTopologyAdapter datastore interactions
RemoteDeviceId holds a reference to device topology, which should be
managed by the user before an adapter is ever instantiated for that
device.

Reduce implicit coupling to default topology by requiring the source of
RemoteDeviceId to properly initialize the operational datastore, such
that it ensures the referenced topology actually exists.

This reduces the overhead for each adapter by two fields, as they are
absolutely superfluous when proper lifecycle is present.

// FIXME: audit callhome-provider to see if it actually does this

JIRA: NETCONF-918
Change-Id: Ib641cc65861bf25fe34ea2ef313ff2a1843e0f97
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>