Remove no-schema reconnect
AbstractNetconfTopology has all the tools it needs to react to device's
failure to establish context, yet we use bonkers routing to the config
datastore and tickle reconnect there.

Remove this code and issue a warning when an attempt to use is made. At
some point we will reimplement this correctly.

JIRA: NETCONF-925
Change-Id: Iaf52f268d7e12f60fd2300d345fbc8c9d30a4c21
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>