WIP: Eliminate LockChangeListener
Configuration of netconf-sal-provider should not directly depend on the
datastore. Move the single leaf from netconf-node-optional into
odl-netconf-device:datastore-access grouping and let users propagate it
via normal configuration updates.

This removes a race as well as explicitly deals with lifecycle of the
device when the configuration is changed.

// FIXME: actually propagate DatastoreAccess

JIRA: NETCONF-921
Change-Id: I07ba946bdef56dfc9ba958384188d93a32e2d974
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>