Rename netconf.api.ModifyAction to EffectiveOperation
This enum has a rather misleading name, as it really is talking about
operations in two contexts -- either an actual operation or the set
default operation. Combined this amounts to the effective operation on
a DataTree.

Let us call it EffectiveOperation, as it ends up being the combination
of default-operation and a potentially-absent explicit operation.

JIRA: NETCONF-932
Change-Id: I1dcf0ffd54fdcf7c7cc853ef595e1a04e69a3147
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>