Rename MountInstance to NetconfDeviceMount
We have a rather useless indirection of components via
NetconfDeviceSalProvider. Eliminate it and promote the only part we care
about to a top-level construct.

Even this may prove unnecessary, as proper lifecycle would just track
the mount point within NetconfDeviceSalFacade. That requires more work
though, so we correctly integrate with netconf-topology-singleton, which
is doing its own thing.

JIRA: NETCONF-918
Change-Id: I014fe4953d4f832b8a9ad43cec95d3d961949716
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>