Convert yanglib to OSGi DS
This is a rather simple component, convert it to OSGi DS, along with
configuration in Config Admin. This means we are removing an initial
configuration XML and replacing it with defaults in
ObjectClassDefinition.

Further changes entail:
- renaming WebInitialized to JaxRsYangLib to make it clear the component
  is a JAX-RS specific thing
- fixing up pom.xml dependency declarations to eliminate maven warnings

Finally we update the documetation to make sure we do not have stale
references.

JIRA: NETCONF-772
Change-Id: I1ef32510c7acf134d8d548003fc607e2800d235e
Signed-off-by: OleksandrZharov <Oleksandr.Zharov@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>