Fix retrieving operation resource
The problem was that it not possible to retrieve operation resource, for example when invoked:
`GET: rests/operations/aaa-cert-rpc:getODLCertificate`
The response is error 400 with message.
According to RFC 8040, p. 22an empty leaf should be returned.

Were created two bodyFor() overloaded methods - which contain
and does not contain identifier as parameter.

getOperations() method was deleted.

JIRA: NETCONF-929
Change-Id: Ic8f0ca5b07e060d36747a110f9049d3d9a2b96c8
Signed-off-by: Yaroslav Lastivka <yaroslav.lastivka@pantheon.tech>