Remove misleading identifier check
Check for presence of mount prefix in request for operation
is misleading. Its fully legal to ask for particular operation
on controller or device.

Remove this check which allows us to invoke requests for particular
operation on controller.

For now the request gives us all operations available on
controller/device and that will be fixed as part of NETCONF-928.

JIRA: NETCONF-929
Change-Id: Ia45ec624640b7b75416071f7675088c2bdab3091
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>