Add yanglib API for non-revision model
The Yanglib API previously used a single API for both
models with revisions and those without. As a result,
it was necessary to append a '/' at the end of the URL request
for non-revision models.

This issue has now been resolved by introducing a separate API
specifically for non-revision models.

Added tests to validate the handling of both revision
and non-revision models. Additionally, the existing Yanglib
tests have been reworked.

JIRA: NETCONF-968
Change-Id: I3a22de90457bd1b2dbb77f3228bcce83a97ba4e1
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>