WIP: Use JsonGenerator in DefinitionGenerator class
Refactor the DefinitionGenerator class to use JsonGenerator
wherever possible.

In the Components class, the schema is currently stored
as an ObjectNode class. To align with the OpenApi specifications,
replace this with a Map<String, ObjectNode>.
This change will ensure consistency with the standards and improve
the code's readability.

JIRA: NETCONF-938
Change-Id: I6badf92d62226faa16383ee6224aa1cdf89fd940
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>