Fix choice cases usage
The way we are currently generating examples for our API documentation
is incorrect. We are including all cases from the "choice" section
in the example payload, but we can only use one case at a time.

Using the default (or first one if not specified) case node in
apidoc example was implemented.

JIRA: NETCONF-983
Change-Id: Ifc32ee00e0de265b4f7e612adb5b8d3783b696a2
Signed-off-by: Yaroslav Lastivka <yaroslav.lastivka@pantheon.tech>
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>